requirements-vercel.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.1
pydantic==2.10.6
semantic-kernel>=1.30.0
openai>=1.61.1
pinecone>=5.4.2
